mode to double

+++ lib/src/models/track_meta.dart
@@ -75,7 +75,7 @@ class TrackMeta extends Object {
 
   ///
   @JsonKey(name: 'mode')
-  int? mode;
+  double? mode;
 
   ///
   @JsonKey(name: 'mode_confidence')
